requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
SQLAlchemy==1.4.49
pydantic==1.10.11
alembic==1.11.1      
async-timeout==4.0.2
httpx==0.28.1
python-dotenv==1.0.0
